requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.6.2
discord.py==2.3.2
frozenlist==1.4.1
h11==0.14.0
idna==3.7
multidict==6.0.5
outcome==1.3.0.post0
PySocks==1.7.1
python-dotenv==1.0.1
selenium==4.21.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.1
urllib3==2.2.1
wsproto==1.2.0
yarl==1.9.4
